image: use float instead of double to store the image timestamp
The image timestamp is uploaded as a float to the GPU. This allows to do
the double->float conversion earlier (when the image is mapped) instead
of doing it just before uploading the relevant uniform.

+++ libnodegl/image.h
@@ -70,7 +70,7 @@ struct image {
     NGLI_ALIGNED_MAT(color_matrix);
     /* mutable fields after initialization */
     NGLI_ALIGNED_MAT(coordinates_matrix);
-    double ts;
+    float ts;
 };
 
 void ngli_image_init(struct image *s, const struct image_params *params, struct texture **planes);
